lots of changes trying to get gfe to run though logic flow with shared functions as much as possible
