+++ spinn_front_end_common/utilities/report_functions/front_end_common_reports.xml
@@ -11,14 +11,6 @@
                 <param_name>processor_to_app_data_base_address</param_name>
                 <param_type>ProcessorToAppDataBaseAddress</param_type>
             </parameter>
-            <parameter>
-                <param_name>dse_on_host</param_name>
-                <param_type>DSEOnHost</param_type>
-            </parameter>
-            <parameter>
-                <param_name>dse_on_chip</param_name>
-                <param_type>DSEOnChip</param_type>
-            </parameter>
         </required_inputs>
         <produces_outputs>
         </produces_outputs>
